+++ Plan/common/src/test/java/com/djrapitops/plan/delivery/webserver/HttpAccessControlTest.java
@@ -70,7 +70,7 @@ class HttpAccessControlTest {
                 ADDRESS,
                 TestConstants.VERSION)));
 
-        Caller caller = system.getExtensionService().register(new ExtensionsDatabaseTest.PlayerExtension())
+        Caller caller = system.getApiServices().getExtensionService().register(new ExtensionsDatabaseTest.PlayerExtension())
                 .orElseThrow(AssertionError::new);
         caller.updatePlayerData(TestConstants.PLAYER_ONE_UUID, TestConstants.PLAYER_ONE_NAME);
